Move `test_getcoor.py` and `test_model.py` and Create `tests/data/`
`tests/data/` is a folder for keeping sample test file
- `nan.geojson` is a test file with `NaN` value and have enough rows for training data
- `normal.geojson` is a test file without `NaN` value and have enough rows for training data
- `small.geojson` is a test file without `NaN` value but don't have enough rows for training data

Move `test_getcoor.py` and `test_model.py` for package consistency
